@@ -1,7 +1,6 @@
 # frozen_string_literal: true
 
 module ApplicationCable
-  # Base channel class for app
-  class Channel < ActionCable::Channel::Base
+  class Channel < ActionCable::Channel::Base #:nodoc:
   end
 end
