@@ -1,7 +1,6 @@
 # frozen_string_literal: true
 
 module ApplicationCable
-  # Base connection class for app
-  class Connection < ActionCable::Connection::Base
+  class Connection < ActionCable::Connection::Base #:nodoc:
   end
 end
