@@ -1,5 +1,4 @@
 # frozen_string_literal: true
 
-# Helpers for app
-module ApplicationHelper
+module ApplicationHelper #:nodoc:
 end
