@@ -1,5 +1,9 @@
 # frozen_string_literal: true
 
-# Base job class for app
-class ApplicationJob < ActiveJob::Base
+class ApplicationJob < ActiveJob::Base #:nodoc:
+  # Automatically retry jobs that encountered a deadlock
+  # retry_on ActiveRecord::Deadlocked
+
+  # Most jobs are safe to ignore if the underlying records are no longer available
+  # discard_on ActiveJob::DeserializationError
 end
