@@ -4,24 +4,18 @@ require_relative 'boot'
 
 require 'rails/all'
 
-# Require the gems listed in Gemfile, including any gems
-# you've limited to :test, :development, or :production.
+# Require the gems listed in Gemfile, including any gems you've limited to :test, :development, or :production.
 Bundler.require(*Rails.groups)
 
 module UlabWebsite
   class Application < Rails::Application # :nodoc:
     # Initialize configuration defaults for originally generated Rails version.
-    config.load_defaults 5.2
-
-    # Serve static assets
-    config.serve_static_assets = true
+    config.load_defaults 6.0
 
     config.i18n.default_locale = :'en-GB'
 
-    # Settings in config/environments/* take precedence over those specified
-    # here.
-    # Application configuration can go into files in config/initializers
-    # -- all .rb files in that directory are automatically loaded after loading
-    # the framework and any gems in your application.
+    # Settings in config/environments/* take precedence over those specified here.
+    # Application configuration can go into files in config/initializers -- all .rb files in that directory are
+    # automatically loaded after loading the framework and any gems in your application.
   end
 end
